created eos_spiner_common to hold common functionality between SpinerEosDependsRHoSie and SpinerEosDependsRhoT, fixed endif's, +clang format
